+++ packages/components/organisms/f-checkout/test-utils/component-objects/f-checkout.component.js
@@ -84,19 +84,10 @@ class Checkout extends Page {
      */
 
     open (checkout) {
-        let authQueryString = '';
-        let createGuestUrl = '&knob-Create%20Guest%20Url=%2Fcreate-guest.json';
-        const checkoutUrl = checkout.isValid ? `&knob-Get%20Checkout%20Url=%2Fcheckout-${checkout.type}.json` : '';
-        const basketUrl = `&knob-Get%20Basket%20Url=%2Fget-basket-${checkout.type}.json`;
-        const placeOrderUrl = '&knob-Place%20Order%20Url=%2Fplace-order.json';
-        const paymentPageUrlPrefix = '&knob-Payment%20Page%20Url%20Prefix=%23%2Fpay';
+        const serviceType = checkout.isValid ? `&knob-Service%20Type=${checkout.type}` : '&knob-Service%20Type=Invalid%20URL';
+        const isLoggedIn = `&knob-Is%20User%20Logged%20In=${checkout.isAuthenticated}`;
 
-        if (checkout.isAuthenticated) {
-            authQueryString = '&knob-Auth%20token=a';
-            createGuestUrl = '';
-        }
-
-        const url = `checkout-component${checkoutUrl}&knob-Available%20Fulfilment%20Url=%2Fcheckout-available-fulfilment.json${authQueryString}${basketUrl}${createGuestUrl}${placeOrderUrl}${paymentPageUrlPrefix}`;
+        const url = `checkout-component${serviceType}${isLoggedIn}`;
         super.openComponent('organism', url);
     }