+++ packages/components/organisms/f-checkout/src/components/Checkout.vue
@@ -175,32 +175,6 @@ export default {
             default: 1000
         },
 
-        getCheckoutTimeout: {
-            type: Number,
-            required: false,
-            default: 1000
-        },
-
-        createGuestTimeout: {
-            type: Number,
-            default: 1000
-        },
-
-        getBasketTimeout: {
-            type: Number,
-            default: 1000
-        },
-
-        placeOrderTimeout: {
-            type: Number,
-            default: 1000
-        },
-
-        updateCheckoutTimeout: {
-            type: Number,
-            default: 1000
-        },
-
         authToken: {
             type: String,
             default: ''
@@ -214,11 +188,6 @@ export default {
         getAddressUrl: {
             type: String,
             required: true
-        },
-
-        getAddressTimeout: {
-            type: Number,
-            default: 1000
         },
 
         applicationName: {
@@ -385,7 +354,7 @@ export default {
                 await this.updateCheckout({
                     url: this.updateCheckoutUrl,
                     data,
-                    timeout: this.updateCheckoutTimeout
+                    timeout: this.checkoutTimeout
                 });
 
                 await this.submitOrder();
@@ -441,7 +410,7 @@ export default {
             await this.placeOrder({
                 url: this.placeOrderUrl,
                 data,
-                timeout: this.placeOrderTimeout
+                timeout: this.checkoutTimeout
             });
         },
 
@@ -462,7 +431,7 @@ export default {
                     url: this.createGuestUrl,
                     tenant: this.tenant,
                     data: createGuestData,
-                    timeout: this.createGuestTimeout
+                    timeout: this.checkoutTimeout
                 });
 
                 this.$emit(EventNames.CheckoutSetupGuestSuccess);
@@ -485,7 +454,7 @@ export default {
             try {
                 await this.getCheckout({
                     url: this.getCheckoutUrl,
-                    timeout: this.getCheckoutTimeout
+                    timeout: this.checkoutTimeout
                 });
 
                 this.$emit(EventNames.CheckoutGetSuccess);
@@ -511,7 +480,7 @@ export default {
                     url: this.getBasketUrl,
                     tenant: this.tenant,
                     language: this.$i18n.locale,
-                    timeout: this.getBasketTimeout
+                    timeout: this.checkoutTimeout
                 });
 
                 this.$emit(EventNames.CheckoutBasketGetSuccess);
@@ -535,7 +504,7 @@ export default {
             try {
                 await this.getAvailableFulfilment({
                     url: this.checkoutAvailableFulfilmentUrl,
-                    timeout: this.getCheckoutTimeout
+                    timeout: this.checkoutTimeout
                 });
 
                 this.$emit(EventNames.CheckoutAvailableFulfilmentGetSuccess);
@@ -561,7 +530,7 @@ export default {
                     url: this.getAddressUrl,
                     tenant: this.tenant,
                     language: this.$i18n.locale,
-                    timeout: this.getAddressTimeout
+                    timeout: this.checkoutTimeout
                 });
 
                 this.$emit(EventNames.CheckoutAddressGetSuccess);